Thach

+++ web_service/src/main/java/com/codegym/web_service/Controller/employeeController/DistributorController.java
@@ -1,7 +1,9 @@
 package com.codegym.web_service.Controller.employeeController;
 
+import com.codegym.dao.entity.Bill;
 import com.codegym.dao.entity.Distributor;
 import com.codegym.dao.entity.TypeOfDistributor;
+import com.codegym.service.BillService;
 import com.codegym.service.DistributorService;
 import com.codegym.service.TypeOfDistributorService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +13,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -20,6 +23,8 @@ public class DistributorController {
     private TypeOfDistributorService typeOfDistributorService;
     @Autowired
     private DistributorService distributorService;
+    @Autowired
+    private BillService billService;
 
     @GetMapping("/distributor/{id}")
     public ResponseEntity<Distributor> findByID(@PathVariable Integer id) {
@@ -69,15 +74,38 @@ public class DistributorController {
         return new ResponseEntity<>(distributorPage, HttpStatus.OK);
     }
 
+    @GetMapping("/distributor/delete/{id}")
+    public ResponseEntity<String> deleteALl(@PathVariable int id) {
+        System.out.println("id cua no ne" + id);
+        this.distributorService.deleteById(id);
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+
 
     @PostMapping("/distributor/deleteAll")
-    public ResponseEntity<String> deleteALl(@RequestBody int[] deleteIdList) {
-        System.out.println(deleteIdList.length);
+    public ResponseEntity<List<Distributor>> deleteALl(@RequestBody int[] deleteIdList) {
+        List<Bill> bills;
+        List<Distributor> deleteIdListFail = new ArrayList<Distributor>();
+        Distributor distributor;
         for (int item : deleteIdList) {
-            this.distributorService.deleteById(item);
+            distributor = this.distributorService.findByIdToDo(item);
+            if (distributor != null) {
+                bills = this.billService.findAllByIdDistributorAndShippingStatusIsNot(distributor, "đã vận chuyển");
+                if (bills.size() > 0) {
+                    deleteIdListFail.add(distributor);
+                } else {
+                    this.distributorService.deleteById(item);
+                }
+            } else {
+                distributor = this.distributorService.findById(item);
+                deleteIdListFail.add(distributor);
+
+            }
+
         }
-        return new ResponseEntity<String>(HttpStatus.OK);
+        return new ResponseEntity<List<Distributor>>(deleteIdListFail, HttpStatus.OK);
     }
+
     @GetMapping("/distributor/exist/{name}/{id}")
     public ResponseEntity<Distributor> findByIsExistName(@PathVariable String name, @PathVariable int id) {
         Distributor distributor = this.distributorService.isExistDistributor(name, id);
@@ -88,4 +116,34 @@ public class DistributorController {
         }
 
     }
+
+
+//    @GetMapping("/distributor/setSession/{id}")
+//    public void resetStatus(@PathVariable int id){
+//        System.out.println(id);
+//        Date date = new Date(System.currentTimeMillis()+10);
+//        this.distributorService.resetStatusDistributor(id);
+//    }
+
+    @GetMapping("/distributorToDo/{id}/{status}")
+    public ResponseEntity<Distributor> findByIDToDo(@PathVariable int id, @PathVariable int status) {
+        Distributor distributor = this.distributorService.findByIdToDo(id);
+        if (distributor != null) {
+            this.distributorService.setStatusDistributor(id, status);
+            this.distributorService.resetSession(id);
+            return new ResponseEntity<>(distributor, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(null, HttpStatus.OK);
+        }
+
+    }
+
+    @GetMapping("/distributorToDoReset/{id}")
+    public void resetStatusDistributorDefault(@PathVariable int id) {
+        this.distributorService.dropEventById(id);
+        this.distributorService.setStatusDistributor(id, 0);
+
+    }
+
+
 }
